use config.mongodb.{server|port} if commander.{host|dbport} are falsy
Fixes #266 (the issue shows the problem and steps to reproduce).

Currently config.mongodb.server and config.mongodb.port are replaced
unconditionally with commander.host and commander.dbport.
They get set to undefined if the command line arguments are missing.

This change keeps the config.mongodb.* values if the command line
arguments are falsy, which is how config.site.port is already handled.